requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
openai>=1.0.0
